Refactor ServerHub workspace presenter
- migrate to createAssetWorkspacePresenter with declarative header, view, and table metadata\n- extract apps, upgrades, and pricing renderers into dedicated modules\n- npm test\n- QA: ServerHub workspace navigation and actions
